test: split core lint warning suites
