+++ packages/core/src/github/issues.test.ts
@@ -1,71 +1,6 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import type { Octokit } from "@octokit/rest";
-import type Database from "better-sqlite3";
-import {
-  listIssues,
-  getIssue,
-  createIssue,
-  updateIssue,
-  closeIssue,
-  reopenIssue,
-  getComments,
-  addComment,
-  updateComment,
-  deleteComment,
-  reassignIssue,
-} from "./issues.js";
-import { createTestDb } from "../db/test-helpers.js";
-import { addRepo } from "../db/repos.js";
-import { setPriority, getPriority } from "../db/priority.js";
-import { setCached, getCached } from "../db/cache.js";
-
-/* ---------- helpers ---------- */
-
-const RAW_ISSUE = {
-  number: 1,
-  title: "Bug report",
-  body: "Something is broken",
-  state: "open",
-  labels: [{ name: "bug", color: "d73a4a", description: "Bug label" }],
-  user: { login: "alice", avatar_url: "https://avatar.test/alice" },
-  created_at: "2026-01-01T00:00:00Z",
-  updated_at: "2026-01-02T00:00:00Z",
-  closed_at: null,
-  html_url: "https://github.com/owner/repo/issues/1",
-};
-
-const RAW_COMMENT = {
-  id: 100,
-  body: "A comment",
-  user: { login: "bob", avatar_url: "https://avatar.test/bob" },
-  created_at: "2026-01-03T00:00:00Z",
-  updated_at: "2026-01-03T00:00:00Z",
-  html_url: "https://github.com/owner/repo/issues/1#issuecomment-100",
-};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type MockFn = ReturnType<typeof vi.fn<(...args: any[]) => any>>;
-
-function makeOctokit() {
-  const paginate = vi.fn() as MockFn;
-  const get = vi.fn() as MockFn;
-  const create = vi.fn() as MockFn;
-  const update = vi.fn() as MockFn;
-  const listComments = vi.fn() as MockFn;
-  const createComment = vi.fn() as MockFn;
-  const updateComment = vi.fn() as MockFn;
-  const deleteComment = vi.fn() as MockFn;
-  const listForRepo = vi.fn() as MockFn;
-
-  const octokit = {
-    paginate,
-    rest: {
-      issues: { listForRepo, get, create, update, listComments, createComment, updateComment, deleteComment },
-    },
-  } as unknown as Octokit;
-
-  return { octokit, paginate, get, create, update, listComments, createComment, updateComment, deleteComment, listForRepo };
-}
+import { describe, it, expect } from "vitest";
+import { listIssues, getIssue, createIssue, updateIssue, closeIssue, reopenIssue } from "./issues.js";
+import { RAW_ISSUE, makeOctokit } from "./issues-test-helpers.js";
 
 /* ---------- listIssues ---------- */
 
@@ -247,259 +182,5 @@ describe("reopenIssue", () => {
     );
 
     await expect(reopenIssue(octokit, "owner", "repo", 999)).rejects.toThrow("Not Found");
-  });
-});
-
-/* ---------- getComments / addComment ---------- */
-
-describe("getComments", () => {
-  it("returns mapped comments", async () => {
-    const { octokit, paginate } = makeOctokit();
-    paginate.mockResolvedValue([RAW_COMMENT]);
-
-    const comments = await getComments(octokit, "owner", "repo", 1);
-    expect(comments).toHaveLength(1);
-    expect(comments[0].id).toBe(100);
-    expect(comments[0].body).toBe("A comment");
-    expect(comments[0].user?.login).toBe("bob");
-  });
-});
-
-describe("addComment", () => {
-  it("creates a comment and returns mapped result", async () => {
-    const { octokit, createComment } = makeOctokit();
-    createComment.mockResolvedValue({ data: RAW_COMMENT });
-
-    const result = await addComment(octokit, "owner", "repo", 1, "A comment");
-    expect(result.body).toBe("A comment");
-    expect(createComment).toHaveBeenCalledWith({
-      owner: "owner",
-      repo: "repo",
-      issue_number: 1,
-      body: "A comment",
-    });
-  });
-});
-
-/* ---------- updateComment ---------- */
-
-describe("updateComment", () => {
-  it("updates a comment and returns mapped result", async () => {
-    const { octokit, updateComment: updateCommentMock } = makeOctokit();
-    const updatedRaw = { ...RAW_COMMENT, body: "Updated body" };
-    updateCommentMock.mockResolvedValue({ data: updatedRaw });
-
-    const result = await updateComment(octokit, "owner", "repo", 100, "Updated body");
-    expect(result.body).toBe("Updated body");
-    expect(result.id).toBe(100);
-    expect(result.user?.login).toBe("bob");
-    expect(updateCommentMock).toHaveBeenCalledWith({
-      owner: "owner",
-      repo: "repo",
-      comment_id: 100,
-      body: "Updated body",
-    });
-  });
-
-  it("propagates API errors", async () => {
-    const { octokit, updateComment: updateCommentMock } = makeOctokit();
-    updateCommentMock.mockRejectedValue(
-      Object.assign(new Error("Not Found"), { status: 404 }),
-    );
-
-    await expect(updateComment(octokit, "owner", "repo", 999, "body")).rejects.toThrow("Not Found");
-  });
-});
-
-/* ---------- deleteComment ---------- */
-
-describe("deleteComment", () => {
-  it("deletes a comment", async () => {
-    const { octokit, deleteComment: deleteCommentMock } = makeOctokit();
-    deleteCommentMock.mockResolvedValue(undefined);
-
-    await deleteComment(octokit, "owner", "repo", 100);
-    expect(deleteCommentMock).toHaveBeenCalledWith({
-      owner: "owner",
-      repo: "repo",
-      comment_id: 100,
-    });
-  });
-
-  it("propagates API errors", async () => {
-    const { octokit, deleteComment: deleteCommentMock } = makeOctokit();
-    deleteCommentMock.mockRejectedValue(
-      Object.assign(new Error("Not Found"), { status: 404 }),
-    );
-
-    await expect(deleteComment(octokit, "owner", "repo", 999)).rejects.toThrow("Not Found");
-  });
-});
-
-/* ---------- reassignIssue ---------- */
-
-describe("reassignIssue", () => {
-  let db: Database.Database;
-  let oldRepoId: number;
-  let newRepoId: number;
-
-  const RAW_NEW_ISSUE = {
-    ...RAW_ISSUE,
-    number: 42,
-    html_url: "https://github.com/other-owner/other-repo/issues/42",
-  };
-
-  function makeReassignOctokit() {
-    const kit = makeOctokit();
-    // getIssue — fetch old issue
-    kit.get.mockResolvedValue({ data: RAW_ISSUE });
-    // createIssue — create on new repo
-    kit.create.mockResolvedValue({ data: RAW_NEW_ISSUE });
-    // addComment — cross-reference
-    kit.createComment.mockResolvedValue({ data: RAW_COMMENT });
-    // closeIssue — close old
-    kit.update.mockResolvedValue({ data: { ...RAW_ISSUE, state: "closed" } });
-    return kit;
-  }
-
-  beforeEach(() => {
-    db = createTestDb();
-    const oldRepo = addRepo(db, { owner: "owner", name: "repo" });
-    const newRepo = addRepo(db, { owner: "other-owner", name: "other-repo" });
-    oldRepoId = oldRepo.id;
-    newRepoId = newRepo.id;
-  });
-
-  it("happy path: creates new issue, comments, closes old, returns result without cleanupWarning", async () => {
-    const { octokit, create, createComment, update } = makeReassignOctokit();
-
-    const result = await reassignIssue(db, octokit, oldRepoId, 1, newRepoId);
-
-    // Verify new issue created on target repo with correct title/body
-    expect(create).toHaveBeenCalledWith({
-      owner: "other-owner",
-      repo: "other-repo",
-      title: "Bug report",
-      body: "Something is broken",
-      labels: undefined,
-    });
-
-    // Verify cross-reference comment added to old issue
-    expect(createComment).toHaveBeenCalledWith({
-      owner: "owner",
-      repo: "repo",
-      issue_number: 1,
-      body: "Moved to other-owner/other-repo#42",
-    });
-
-    // Verify old issue closed
-    expect(update).toHaveBeenCalledWith({
-      owner: "owner",
-      repo: "repo",
-      issue_number: 1,
-      state: "closed",
-    });
-
-    // Verify return value
-    expect(result).toEqual({
-      newIssueNumber: 42,
-      newIssueUrl: "https://github.com/other-owner/other-repo/issues/42",
-      newOwner: "other-owner",
-      newRepo: "other-repo",
-    });
-    expect(result.cleanupWarning).toBeUndefined();
-  });
-
-  it("throws when old and new repo IDs are the same", async () => {
-    const { octokit } = makeReassignOctokit();
-
-    await expect(
-      reassignIssue(db, octokit, oldRepoId, 1, oldRepoId),
-    ).rejects.toThrow("Cannot re-assign an issue to the same repo");
-  });
-
-  it("throws when old repo ID does not exist in DB", async () => {
-    const { octokit } = makeReassignOctokit();
-
-    await expect(
-      reassignIssue(db, octokit, 9999, 1, newRepoId),
-    ).rejects.toThrow("Old repo (id 9999) not found");
-  });
-
-  it("throws when new repo ID does not exist in DB", async () => {
-    const { octokit } = makeReassignOctokit();
-
-    await expect(
-      reassignIssue(db, octokit, oldRepoId, 1, 9999),
-    ).rejects.toThrow("New repo (id 9999) not found");
-  });
-
-  it("returns result with cleanupWarning when closeIssue fails after create succeeds", async () => {
-    const { octokit, update } = makeReassignOctokit();
-
-    // addComment succeeds (uses createComment), but closeIssue (uses update) fails.
-    update.mockRejectedValue(new Error("API rate limit exceeded"));
-
-    const result = await reassignIssue(db, octokit, oldRepoId, 1, newRepoId);
-
-    // Should NOT throw
-    expect(result.newIssueNumber).toBe(42);
-    expect(result.newIssueUrl).toBe("https://github.com/other-owner/other-repo/issues/42");
-    expect(result.cleanupWarning).toContain("could not be closed");
-    expect(result.cleanupWarning).toContain("API rate limit exceeded");
-  });
-
-  it("returns result with cleanupWarning when addComment fails after create succeeds", async () => {
-    const { octokit, createComment } = makeReassignOctokit();
-
-    createComment.mockRejectedValue(new Error("Forbidden"));
-
-    const result = await reassignIssue(db, octokit, oldRepoId, 1, newRepoId);
-
-    expect(result.newIssueNumber).toBe(42);
-    expect(result.cleanupWarning).toContain("could not be closed");
-    expect(result.cleanupWarning).toContain("Forbidden");
-  });
-
-  it("migrates priority from old repo/issue to new repo/issue", async () => {
-    const { octokit } = makeReassignOctokit();
-
-    // Seed a priority on the old repo/issue
-    setPriority(db, oldRepoId, 1, "high");
-
-    await reassignIssue(db, octokit, oldRepoId, 1, newRepoId);
-
-    // New repo/issue should have the priority
-    expect(getPriority(db, newRepoId, 42)).toBe("high");
-    // Old repo/issue priority should be deleted (returns default "normal")
-    expect(getPriority(db, oldRepoId, 1)).toBe("normal");
-  });
-
-  it("clears all 5 cache keys after reassignment", async () => {
-    const { octokit } = makeReassignOctokit();
-
-    // Seed cache entries for all 5 keys that reassignIssue should clear
-    const cacheKeys = [
-      "issues:owner/repo",
-      "issue-detail:owner/repo#1",
-      "issue-header:owner/repo#1",
-      "issue-content:owner/repo#1",
-      "issues:other-owner/other-repo",
-    ];
-    for (const key of cacheKeys) {
-      setCached(db, key, { placeholder: true });
-    }
-
-    // Verify they exist before
-    for (const key of cacheKeys) {
-      expect(getCached(db, key)).not.toBeNull();
-    }
-
-    await reassignIssue(db, octokit, oldRepoId, 1, newRepoId);
-
-    // All 5 cache keys should be cleared
-    for (const key of cacheKeys) {
-      expect(getCached(db, key)).toBeNull();
-    }
   });
 });
