+++ packages/core/src/launch/launch-execute-precheck.test.ts
@@ -1,0 +1,184 @@
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import type Database from "better-sqlite3";
+import type { Octokit } from "@octokit/rest";
+import { executeLaunch } from "./launch.js";
+import { createTestDb } from "../db/test-helpers.js";
+import { addRepo } from "../db/repos.js";
+import { recordDeployment } from "../db/deployments.js";
+import * as deploymentsModule from "../db/deployments.js";
+
+// Stub every module side-effect outside the pre-check's reach so the
+// test only measures ordering: did `prepareWorkspace` run even though
+// a live deployment already existed? `vi.hoisted` declares the spy at
+// the same time vi.mock is hoisted so the factory can reference it.
+const { prepareWorkspaceSpy } = vi.hoisted(() => ({
+  prepareWorkspaceSpy: vi.fn(async () => ({
+    path: "/tmp/fake-workspace",
+    branchName: "fake-branch",
+  })),
+}));
+
+vi.mock("./workspace.js", () => ({
+  prepareWorkspace: prepareWorkspaceSpy,
+}));
+
+const { verifyTtydSpy, spawnTtydSpy, allocatePortSpy } = vi.hoisted(() => ({
+  verifyTtydSpy: vi.fn(),
+  spawnTtydSpy: vi.fn(async () => ({ pid: 12345, port: 7700 })),
+  allocatePortSpy: vi.fn(async () => 7700),
+}));
+
+vi.mock("./ttyd.js", () => ({
+  verifyTtyd: verifyTtydSpy,
+  spawnTtyd: spawnTtydSpy,
+  allocatePort: allocatePortSpy,
+  tmuxSessionName: (repo: string, issueNumber: number) =>
+    `issuectl-${repo}-${issueNumber}`,
+}));
+
+const { reserveTtydPortSpy, updateTtydInfoSpy } = vi.hoisted(() => ({
+  reserveTtydPortSpy: vi.fn(),
+  updateTtydInfoSpy: vi.fn(),
+}));
+
+vi.mock("../db/deployments.js", async () => {
+  const actual =
+    await vi.importActual<typeof import("../db/deployments.js")>(
+      "../db/deployments.js",
+    );
+  return {
+    ...actual,
+    reserveTtydPort: reserveTtydPortSpy,
+    updateTtydInfo: updateTtydInfoSpy,
+  };
+});
+
+vi.mock("../data/issues.js", () => ({
+  getIssueDetail: async () => ({
+    issue: {
+      number: 42,
+      title: "test issue",
+      body: "",
+      state: "open",
+      labels: [],
+      user: null,
+      commentCount: 0,
+      createdAt: "2026-04-12T00:00:00Z",
+      updatedAt: "2026-04-12T00:00:00Z",
+      closedAt: null,
+      htmlUrl: "https://example.invalid",
+    },
+    comments: [],
+    referencedFiles: [],
+  }),
+}));
+
+vi.mock("./context.js", () => ({
+  assembleContext: () => "fake context",
+  writeContextFile: async () => "/tmp/fake-context.md",
+}));
+
+vi.mock("../github/labels.js", async () => {
+  const actual =
+    await vi.importActual<typeof import("../github/labels.js")>(
+      "../github/labels.js",
+    );
+  return {
+    ...actual,
+    ensureLifecycleLabels: async () => {},
+    addLabel: async () => {},
+    addLabels: async () => {},
+    removeLabel: async () => {},
+  };
+});
+
+describe("executeLaunch duplicate-deployment pre-check", () => {
+  let db: Database.Database;
+
+  beforeEach(() => {
+    prepareWorkspaceSpy.mockClear();
+    verifyTtydSpy.mockClear();
+    spawnTtydSpy.mockClear();
+    allocatePortSpy.mockClear();
+    reserveTtydPortSpy.mockClear();
+    updateTtydInfoSpy.mockClear();
+    db = createTestDb();
+  });
+
+  it("refuses to launch and skips workspace prep when a live deployment already exists", async () => {
+    const repo = addRepo(db, {
+      owner: "acme",
+      name: "api",
+      localPath: "/tmp/fake",
+    });
+    recordDeployment(db, {
+      repoId: repo.id,
+      issueNumber: 42,
+      branchName: "prior-branch",
+      workspaceMode: "existing",
+      workspacePath: "/tmp/prior-workspace",
+    });
+
+    await expect(
+      executeLaunch(db, {} as Octokit, {
+        owner: "acme",
+        repo: "api",
+        issueNumber: 42,
+        branchName: "new-branch",
+        workspaceMode: "existing",
+        selectedComments: [],
+        selectedFiles: [],
+      }),
+    ).rejects.toThrow(/already has an active deployment/);
+
+    expect(prepareWorkspaceSpy).not.toHaveBeenCalled();
+  });
+
+  it("translates a race-path UNIQUE violation into the friendly duplicate-launch error", async () => {
+    // Simulate the race: the pre-check returns false (the concurrent
+    // live row was inserted between the SELECT and the INSERT), but
+    // the real live row already exists in the DB, so recordDeployment
+    // trips idx_deployments_live. The catch at launch.ts step 8 must
+    // translate the raw SqliteError into the same friendly message
+    // the pre-check throws.
+    const repo = addRepo(db, {
+      owner: "acme",
+      name: "api",
+      localPath: "/tmp/fake",
+    });
+    recordDeployment(db, {
+      repoId: repo.id,
+      issueNumber: 42,
+      branchName: "winner-branch",
+      workspaceMode: "existing",
+      workspacePath: "/tmp/winner",
+    });
+
+    const preCheckSpy = vi
+      .spyOn(deploymentsModule, "hasLiveDeploymentForIssue")
+      .mockReturnValue(false);
+
+    try {
+      await expect(
+        executeLaunch(db, {} as Octokit, {
+          owner: "acme",
+          repo: "api",
+          issueNumber: 42,
+          branchName: "loser-branch",
+          workspaceMode: "existing",
+          selectedComments: [],
+          selectedFiles: [],
+        }),
+      ).rejects.toThrow(/already has an active deployment/);
+
+      // prepareWorkspace *should* have run — this is the race path,
+      // not the pre-check path. We're asserting the catch works once
+      // we've gotten past the optimistic check.
+      expect(prepareWorkspaceSpy).toHaveBeenCalledTimes(1);
+    } finally {
+      preCheckSpy.mockRestore();
+    }
+  });
+
+
+});
