+++ packages/core/src/launch/ttyd.test.ts
@@ -1,5 +1,4 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import type Database from "better-sqlite3";
 
 /* ------------------------------------------------------------------ */
 /*  Mocks                                                              */
@@ -26,53 +25,7 @@ vi.mock("node:net", () => ({
   default: { connect: netConnectSpy },
 }));
 
-// Import after mocks are in place.
-import {
-  verifyTtyd,
-  killTtyd,
-  isTtydAlive,
-  isTmuxSessionAlive,
-  allocatePort,
-  spawnTtyd,
-  respawnTtyd,
-  reconcileOrphanedDeployments,
-  tmuxSessionName,
-} from "./ttyd.js";
-
-/* ------------------------------------------------------------------ */
-/*  Helpers                                                            */
-/* ------------------------------------------------------------------ */
-
-/** Minimal fake DB that satisfies the functions under test. */
-function makeFakeDb(rows: Record<string, unknown>[] = []): Database.Database {
-  return {
-    prepare: vi.fn(() => ({
-      all: vi.fn(() => rows),
-      run: vi.fn(),
-    })),
-  } as unknown as Database.Database;
-}
-
-/** Create an EventEmitter-like fake socket for net.connect mocking. */
-function fakeSocket(behavior: "connect" | "error") {
-  const handlers: Record<string, (() => void)[]> = {};
-  const socket = {
-    on(event: string, cb: () => void) {
-      (handlers[event] ??= []).push(cb);
-      return socket;
-    },
-    destroy: vi.fn(),
-  };
-  // Fire the appropriate event on the next microtask.
-  queueMicrotask(() => {
-    for (const cb of handlers[behavior] ?? []) cb();
-  });
-  return socket;
-}
-
-/* ------------------------------------------------------------------ */
-/*  tmuxSessionName                                                    */
-/* ------------------------------------------------------------------ */
+import { verifyTtyd, killTtyd, isTtydAlive, tmuxSessionName } from "./ttyd.js";
 
 describe("tmuxSessionName", () => {
   it("produces a predictable name from repo and issue number", () => {
@@ -278,644 +231,5 @@ describe("isTtydAlive", () => {
     });
     expect(() => isTtydAlive(1)).toThrow("EINVAL");
     killSpy.mockRestore();
-  });
-});
-
-/* ------------------------------------------------------------------ */
-/*  isTmuxSessionAlive                                                 */
-/* ------------------------------------------------------------------ */
-
-describe("isTmuxSessionAlive", () => {
-  beforeEach(() => {
-    execFileSyncSpy.mockReset();
-  });
-
-  it("returns true when tmux session exists (exit code 0)", () => {
-    execFileSyncSpy.mockReturnValue(Buffer.from(""));
-    expect(isTmuxSessionAlive("issuectl-repo-42")).toBe(true);
-    expect(execFileSyncSpy).toHaveBeenCalledWith(
-      "tmux", ["has-session", "-t", "issuectl-repo-42"],
-      expect.objectContaining({ stdio: "ignore", timeout: 10_000 }),
-    );
-  });
-
-  it("returns false when tmux session does not exist (exit code 1)", () => {
-    execFileSyncSpy.mockImplementation(() => {
-      throw Object.assign(new Error("session not found"), { status: 1 });
-    });
-    expect(isTmuxSessionAlive("issuectl-repo-42")).toBe(false);
-  });
-
-  it("returns false when tmux is not installed (ENOENT)", () => {
-    execFileSyncSpy.mockImplementation(() => {
-      throw Object.assign(new Error("not found"), { code: "ENOENT" });
-    });
-    expect(isTmuxSessionAlive("issuectl-repo-42")).toBe(false);
-  });
-
-  it("throws on unexpected errors (ETIMEDOUT) to prevent false 'dead' cascades", () => {
-    execFileSyncSpy.mockImplementation(() => {
-      throw Object.assign(new Error("timed out"), { code: "ETIMEDOUT" });
-    });
-    expect(() => isTmuxSessionAlive("issuectl-repo-42")).toThrow(
-      'tmux has-session failed unexpectedly for "issuectl-repo-42"',
-    );
-  });
-
-  it("throws on unexpected errors (EPERM) to prevent false 'dead' cascades", () => {
-    execFileSyncSpy.mockImplementation(() => {
-      throw Object.assign(new Error("EPERM"), { code: "EPERM" });
-    });
-    expect(() => isTmuxSessionAlive("issuectl-repo-42")).toThrow(
-      "tmux has-session failed unexpectedly",
-    );
-  });
-
-  it("preserves original error as cause when throwing", () => {
-    const original = Object.assign(new Error("timed out"), { code: "ETIMEDOUT" });
-    execFileSyncSpy.mockImplementation(() => {
-      throw original;
-    });
-    try {
-      isTmuxSessionAlive("issuectl-repo-42");
-    } catch (err) {
-      expect((err as Error).cause).toBe(original);
-    }
-  });
-});
-
-/* ------------------------------------------------------------------ */
-/*  allocatePort                                                       */
-/* ------------------------------------------------------------------ */
-
-describe("allocatePort", () => {
-  beforeEach(() => {
-    netConnectSpy.mockReset();
-  });
-
-  it("returns 7700 when no ports are in use", async () => {
-    const db = makeFakeDb([]);
-    // All probes fail → port is free.
-    netConnectSpy.mockImplementation(() => fakeSocket("error"));
-
-    const port = await allocatePort(db);
-    expect(port).toBe(7700);
-  });
-
-  it("skips ports used by active deployments", async () => {
-    const db = makeFakeDb([{ ttyd_port: 7700 }, { ttyd_port: 7701 }]);
-    netConnectSpy.mockImplementation(() => fakeSocket("error"));
-
-    const port = await allocatePort(db);
-    expect(port).toBe(7702);
-  });
-
-  it("skips ports that are in use (TCP probe succeeds)", async () => {
-    const db = makeFakeDb([]);
-    // First probe succeeds (port in use), rest fail.
-    let callCount = 0;
-    netConnectSpy.mockImplementation(() => {
-      callCount++;
-      return fakeSocket(callCount === 1 ? "connect" : "error");
-    });
-
-    const port = await allocatePort(db);
-    expect(port).toBe(7701);
-  });
-
-  it("throws when all 100 ports are taken", async () => {
-    // All ports claimed in DB.
-    const allPorts = Array.from({ length: 100 }, (_, i) => ({
-      ttyd_port: 7700 + i,
-    }));
-    const db = makeFakeDb(allPorts);
-
-    await expect(allocatePort(db)).rejects.toThrow(
-      "All ttyd ports (7700–7799) are in use",
-    );
-  });
-});
-
-/* ------------------------------------------------------------------ */
-/*  spawnTtyd                                                          */
-/* ------------------------------------------------------------------ */
-
-describe("spawnTtyd", () => {
-  beforeEach(() => {
-    spawnSpy.mockReset();
-    execFileSyncSpy.mockReset();
-    // execFileSync is used for tmux calls in spawnTtyd — default to no-op
-    execFileSyncSpy.mockReturnValue(Buffer.from(""));
-  });
-
-  it("spawns ttyd with correct arguments and returns PID + port", async () => {
-    const unrefSpy = vi.fn();
-    spawnSpy.mockReturnValue({ pid: 42, unref: unrefSpy, on: vi.fn() });
-    // Health check — isTtydAlive calls process.kill(pid, 0)
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    const result = await spawnTtyd({
-      port: 7700,
-      workspacePath: "/home/user/project",
-      contextFilePath: "/tmp/ctx.md",
-      agentCommand: "claude --dangerously-skip-permissions",
-      sessionName: "issuectl-myrepo-42",
-    });
-
-    expect(result).toEqual({ pid: 42, port: 7700 });
-    expect(unrefSpy).toHaveBeenCalled();
-
-    // tmux session should be created first via execFileSync
-    const tmuxCall = execFileSyncSpy.mock.calls.find(
-      (c: unknown[]) => c[0] === "tmux" && (c[1] as string[])[0] === "new-session",
-    )!;
-    expect(tmuxCall).toBeDefined();
-    const tmuxArgs = tmuxCall[1] as string[];
-    expect(tmuxArgs.slice(0, 8)).toEqual([
-      "new-session", "-d", "-x", "40", "-y", "24", "-s", "issuectl-myrepo-42",
-    ]);
-    // The shell command passed to tmux redirects the context into the agent
-    const tmuxCmd = tmuxArgs[8];
-    expect(tmuxCmd).toContain("bash -lic");
-    expect(tmuxCmd).toContain("/home/user/project");
-    expect(tmuxCmd).toContain("/tmp/ctx.md");
-    expect(tmuxCmd).toContain("claude --dangerously-skip-permissions");
-    expect(tmuxCmd).toContain("< ");
-    expect(tmuxCmd).toContain("; exit");
-
-    // tmux session options (with timeout)
-    expect(execFileSyncSpy).toHaveBeenCalledWith("tmux", [
-      "set-option", "-t", "issuectl-myrepo-42", "status", "off",
-    ], { timeout: 10_000 });
-    expect(execFileSyncSpy).toHaveBeenCalledWith("tmux", [
-      "set-option", "-t", "issuectl-myrepo-42", "window-size", "largest",
-    ], { timeout: 10_000 });
-
-    // ttyd serves tmux attach (not bash -lic)
-    const [bin, args, opts] = spawnSpy.mock.calls[0] as [string, string[], Record<string, unknown>];
-    expect(bin).toBe("ttyd");
-    expect(args).toEqual([
-      "-W", "-i", "127.0.0.1", "-p", "7700", "-q",
-      "tmux", "attach-session", "-t", "issuectl-myrepo-42",
-    ]);
-    expect(opts).toEqual({ detached: true, stdio: "ignore" });
-    killSpy.mockRestore();
-  });
-
-  it("passes context as an argument for interactive Codex launches", async () => {
-    spawnSpy.mockReturnValue({ pid: 42, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await spawnTtyd({
-      port: 7700,
-      workspacePath: "/home/user/project",
-      contextFilePath: "/tmp/ctx.md",
-      agentCommand: "codex --sandbox danger-full-access --ask-for-approval never",
-      agentInputMode: "argument",
-      sessionName: "issuectl-myrepo-43",
-    });
-
-    const tmuxCmd = execFileSyncSpy.mock.calls.find(
-      (c: unknown[]) => c[0] === "tmux" && (c[1] as string[])[0] === "new-session",
-    )![1][8] as string;
-    expect(tmuxCmd).toContain("codex --sandbox danger-full-access --ask-for-approval never");
-    expect(tmuxCmd).toContain("$(cat ");
-    expect(tmuxCmd).toContain("/tmp/ctx.md");
-    expect(tmuxCmd).not.toContain(" | codex");
-    killSpy.mockRestore();
-  });
-
-  it("binds ttyd to loopback interface (-i 127.0.0.1)", async () => {
-    spawnSpy.mockReturnValue({ pid: 42, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await spawnTtyd({
-      port: 7700,
-      workspacePath: "/tmp/ws",
-      contextFilePath: "/tmp/ctx.md",
-      agentCommand: "claude",
-      sessionName: "issuectl-test-1",
-    });
-
-    const args = (spawnSpy.mock.calls[0] as [string, string[]])[1];
-    expect(args).toContain("-i");
-    expect(args).toContain("127.0.0.1");
-    const iIdx = args.indexOf("-i");
-    expect(args[iIdx + 1]).toBe("127.0.0.1");
-    expect(iIdx).toBeLessThan(args.indexOf("-p"));
-    killSpy.mockRestore();
-  });
-
-  it("escapes paths with single quotes", async () => {
-    spawnSpy.mockReturnValue({ pid: 1, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await spawnTtyd({
-      port: 7700,
-      workspacePath: "/home/user/it's a project",
-      contextFilePath: "/tmp/file.md",
-      agentCommand: "claude",
-      sessionName: "issuectl-test-2",
-    });
-
-    // The escaped path ends up in the tmux new-session command string.
-    // It goes through shellEscape twice: once for the inner command,
-    // once when wrapping in bash -lic, so the quote escaping is doubled.
-    const tmuxCmd = execFileSyncSpy.mock.calls.find(
-      (c: unknown[]) => c[0] === "tmux" && (c[1] as string[])[0] === "new-session",
-    )![1][8] as string;
-    expect(tmuxCmd).toContain("it");
-    expect(tmuxCmd).toContain("s a project");
-    killSpy.mockRestore();
-  });
-
-  it("creates tmux session before spawning ttyd", async () => {
-    const callOrder: string[] = [];
-    execFileSyncSpy.mockImplementation((cmd: string) => {
-      if (cmd === "tmux") callOrder.push("tmux");
-      return Buffer.from("");
-    });
-    spawnSpy.mockImplementation(() => {
-      callOrder.push("spawn");
-      return { pid: 1, unref: vi.fn(), on: vi.fn() };
-    });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await spawnTtyd({
-      port: 7700,
-      workspacePath: "/tmp",
-      contextFilePath: "/tmp/ctx.md",
-      agentCommand: "claude",
-      sessionName: "issuectl-test-order",
-    });
-
-    // All 3 tmux calls (new-session, set status, set window-size) must
-    // happen before the ttyd spawn so clients can attach immediately.
-    expect(callOrder).toEqual(["tmux", "tmux", "tmux", "spawn"]);
-    killSpy.mockRestore();
-  });
-
-  it("propagates tmux new-session failure", async () => {
-    execFileSyncSpy.mockImplementation((cmd: string, args: string[]) => {
-      if (cmd === "tmux" && args[0] === "new-session") {
-        throw new Error("duplicate session: issuectl-test-dup");
-      }
-      return Buffer.from("");
-    });
-
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-test-dup",
-      }),
-    ).rejects.toThrow("duplicate session: issuectl-test-dup");
-
-    // ttyd should NOT have been spawned
-    expect(spawnSpy).not.toHaveBeenCalled();
-  });
-
-  it("uses session name in tmux attach command for ttyd", async () => {
-    spawnSpy.mockReturnValue({ pid: 1, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await spawnTtyd({
-      port: 7700,
-      workspacePath: "/tmp",
-      contextFilePath: "/tmp/ctx.md",
-      agentCommand: "claude",
-      sessionName: "issuectl-special-chars-99",
-    });
-
-    const args = (spawnSpy.mock.calls[0] as [string, string[]])[1];
-    // The last 3 args should be the tmux attach command
-    expect(args.slice(-3)).toEqual([
-      "attach-session", "-t", "issuectl-special-chars-99",
-    ]);
-    killSpy.mockRestore();
-  });
-
-  it("rejects invalid session names containing dots or colons", async () => {
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-my.project-42",
-      }),
-    ).rejects.toThrow("Invalid tmux session name");
-
-    expect(execFileSyncSpy).not.toHaveBeenCalledWith(
-      "tmux", expect.arrayContaining(["new-session"]), expect.anything(),
-    );
-  });
-
-  it("cleans up tmux session when set-option fails", async () => {
-    const tmuxCalls: string[][] = [];
-    execFileSyncSpy.mockImplementation((cmd: string, args: string[]) => {
-      if (cmd === "tmux") {
-        tmuxCalls.push(args);
-        if (args[0] === "set-option" && args[3] === "status") {
-          throw new Error("tmux set-option failed");
-        }
-      }
-      return Buffer.from("");
-    });
-
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-test-cleanup",
-      }),
-    ).rejects.toThrow("tmux set-option failed");
-
-    // tmux kill-session should have been called for cleanup
-    const killCall = tmuxCalls.find((args) => args[0] === "kill-session");
-    expect(killCall).toBeDefined();
-    expect(killCall![2]).toBe("issuectl-test-cleanup");
-    expect(spawnSpy).not.toHaveBeenCalled();
-  });
-
-  it("cleans up tmux session when health check fails (ttyd dies)", async () => {
-    spawnSpy.mockReturnValue({ pid: 99, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => {
-      throw Object.assign(new Error("ESRCH"), { code: "ESRCH" });
-    });
-
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-test-health",
-      }),
-    ).rejects.toThrow("ttyd process 99 died immediately after spawn");
-
-    // tmux kill-session should have been called for cleanup
-    expect(execFileSyncSpy).toHaveBeenCalledWith(
-      "tmux", ["kill-session", "-t", "issuectl-test-health"],
-      expect.objectContaining({ stdio: "ignore" }),
-    );
-    killSpy.mockRestore();
-  });
-
-  it("throws when no PID is returned", async () => {
-    spawnSpy.mockReturnValue({ pid: undefined, unref: vi.fn(), on: vi.fn() });
-
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-test-3",
-      }),
-    ).rejects.toThrow("Failed to spawn ttyd: no PID returned");
-  });
-
-  it("throws when process dies immediately after spawn", async () => {
-    spawnSpy.mockReturnValue({ pid: 99, unref: vi.fn(), on: vi.fn() });
-    // Health check — process is dead
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => {
-      throw Object.assign(new Error("ESRCH"), { code: "ESRCH" });
-    });
-
-    await expect(
-      spawnTtyd({
-        port: 7700,
-        workspacePath: "/tmp",
-        contextFilePath: "/tmp/ctx.md",
-        agentCommand: "claude",
-        sessionName: "issuectl-test-4",
-      }),
-    ).rejects.toThrow(
-      "ttyd process 99 died immediately after spawn",
-    );
-    killSpy.mockRestore();
-  });
-});
-
-/* ------------------------------------------------------------------ */
-/*  respawnTtyd                                                        */
-/* ------------------------------------------------------------------ */
-
-describe("respawnTtyd", () => {
-  beforeEach(() => {
-    spawnSpy.mockReset();
-    execFileSyncSpy.mockReset();
-  });
-
-  it("spawns ttyd against existing tmux session and returns new PID", async () => {
-    const unrefSpy = vi.fn();
-    spawnSpy.mockReturnValue({ pid: 88, unref: unrefSpy, on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    const result = await respawnTtyd(7700, "issuectl-repo-42");
-
-    expect(result).toEqual({ pid: 88 });
-    expect(unrefSpy).toHaveBeenCalled();
-
-    const [bin, args, opts] = spawnSpy.mock.calls[0] as [string, string[], Record<string, unknown>];
-    expect(bin).toBe("ttyd");
-    expect(args).toEqual([
-      "-W", "-i", "127.0.0.1", "-p", "7700", "-q",
-      "tmux", "attach-session", "-t", "issuectl-repo-42",
-    ]);
-    expect(opts).toEqual({ detached: true, stdio: "ignore" });
-    killSpy.mockRestore();
-  });
-
-  it("does NOT create a new tmux session", async () => {
-    spawnSpy.mockReturnValue({ pid: 88, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => true);
-
-    await respawnTtyd(7700, "issuectl-repo-42");
-
-    // execFileSync should NOT have been called (no tmux new-session)
-    expect(execFileSyncSpy).not.toHaveBeenCalled();
-    killSpy.mockRestore();
-  });
-
-  it("throws when ttyd dies immediately after respawn", async () => {
-    spawnSpy.mockReturnValue({ pid: 99, unref: vi.fn(), on: vi.fn() });
-    const killSpy = vi.spyOn(process, "kill").mockImplementation(() => {
-      throw Object.assign(new Error("ESRCH"), { code: "ESRCH" });
-    });
-
-    await expect(respawnTtyd(7700, "issuectl-repo-42")).rejects.toThrow(
-      "ttyd process 99 died immediately after respawn",
-    );
-    killSpy.mockRestore();
-  });
-
-  it("throws when no PID is returned", async () => {
-    spawnSpy.mockReturnValue({ pid: undefined, unref: vi.fn(), on: vi.fn() });
-
-    await expect(respawnTtyd(7700, "issuectl-repo-42")).rejects.toThrow(
-      "Failed to respawn ttyd: no PID returned",
-    );
-  });
-
-  it("rejects invalid session names containing dots or colons", async () => {
-    await expect(respawnTtyd(7700, "issuectl-my.project-42")).rejects.toThrow(
-      "Invalid tmux session name",
-    );
-
-    // spawn should never have been called
-    expect(spawnSpy).not.toHaveBeenCalled();
-  });
-});
-
-/* ------------------------------------------------------------------ */
-/*  reconcileOrphanedDeployments                                       */
-/* ------------------------------------------------------------------ */
-
-describe("reconcileOrphanedDeployments", () => {
-  beforeEach(() => {
-    execFileSyncSpy.mockReset();
-  });
-
-  it("marks deployments as ended only when tmux session is gone", () => {
-    // Session "issuectl-repoA-10" is alive, "issuectl-repoB-20" is dead.
-    execFileSyncSpy.mockImplementation((_cmd: string, args: string[]) => {
-      if (args[0] === "has-session" && args[2] === "issuectl-repoA-10") {
-        return Buffer.from("");
-      }
-      throw Object.assign(new Error("session not found"), { status: 1 });
-    });
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-
-    const runSpy = vi.fn();
-    const db = {
-      prepare: vi.fn((sql: string) => {
-        if (sql.includes("SELECT")) {
-          return {
-            all: vi.fn(() => [
-              { id: 1, issue_number: 10, repo_name: "repoA" },
-              { id: 2, issue_number: 20, repo_name: "repoB" },
-            ]),
-          };
-        }
-        return { run: runSpy };
-      }),
-    } as unknown as Database.Database;
-
-    reconcileOrphanedDeployments(db);
-
-    // Only deployment 2 should be ended (tmux session gone).
-    expect(runSpy).toHaveBeenCalledTimes(1);
-    expect(runSpy).toHaveBeenCalledWith(2);
-    expect(warnSpy).toHaveBeenCalledWith(
-      expect.stringContaining("Reconciled orphaned deployment 2"),
-    );
-
-    warnSpy.mockRestore();
-  });
-
-  it("does nothing when all tmux sessions are alive", () => {
-    execFileSyncSpy.mockReturnValue(Buffer.from(""));
-
-    const runSpy = vi.fn();
-    const db = {
-      prepare: vi.fn((sql: string) => {
-        if (sql.includes("SELECT")) {
-          return {
-            all: vi.fn(() => [
-              { id: 1, issue_number: 10, repo_name: "repoA" },
-            ]),
-          };
-        }
-        return { run: runSpy };
-      }),
-    } as unknown as Database.Database;
-
-    reconcileOrphanedDeployments(db);
-
-    expect(runSpy).not.toHaveBeenCalled();
-  });
-
-  it("only queries deployments that have a ttyd_pid (excludes pending)", () => {
-    execFileSyncSpy.mockReturnValue(Buffer.from(""));
-
-    const prepareSpy = vi.fn((sql: string) => {
-      if (sql.includes("SELECT")) {
-        return { all: vi.fn(() => []) };
-      }
-      return { run: vi.fn() };
-    });
-    const db = { prepare: prepareSpy } as unknown as Database.Database;
-
-    reconcileOrphanedDeployments(db);
-
-    const selectCall = prepareSpy.mock.calls.find(
-      (c: unknown[]) => (c[0] as string).includes("SELECT"),
-    );
-    expect(selectCall).toBeDefined();
-    expect(selectCall![0]).toContain("ttyd_pid IS NOT NULL");
-  });
-
-  it("logs error and does not throw when query fails", () => {
-    const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const db = {
-      prepare: vi.fn(() => {
-        throw new Error("DB locked");
-      }),
-    } as unknown as Database.Database;
-
-    // Should not throw
-    expect(() => reconcileOrphanedDeployments(db)).not.toThrow();
-    expect(errorSpy).toHaveBeenCalledWith(
-      "[issuectl] Failed to query deployments for reconciliation:",
-      expect.any(Error),
-    );
-    errorSpy.mockRestore();
-  });
-
-  it("continues reconciling other rows when one row fails (per-row isolation)", () => {
-    // Row 1 causes isTmuxSessionAlive to throw (ETIMEDOUT), row 2 is dead.
-    execFileSyncSpy.mockImplementation((_cmd: string, args: string[]) => {
-      if (args[0] === "has-session" && args[2] === "issuectl-repoA-10") {
-        throw Object.assign(new Error("timed out"), { code: "ETIMEDOUT" });
-      }
-      // repoB session is gone (exit code 1)
-      throw Object.assign(new Error("session not found"), { status: 1 });
-    });
-    const errorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-
-    const runSpy = vi.fn();
-    const db = {
-      prepare: vi.fn((sql: string) => {
-        if (sql.includes("SELECT")) {
-          return {
-            all: vi.fn(() => [
-              { id: 1, issue_number: 10, repo_name: "repoA" },
-              { id: 2, issue_number: 20, repo_name: "repoB" },
-            ]),
-          };
-        }
-        return { run: runSpy };
-      }),
-    } as unknown as Database.Database;
-
-    reconcileOrphanedDeployments(db);
-
-    // Row 1 should have failed (logged), row 2 should still be reconciled.
-    expect(errorSpy).toHaveBeenCalledWith(
-      "[issuectl] Failed to reconcile deployment 1:",
-      expect.any(Error),
-    );
-    expect(runSpy).toHaveBeenCalledTimes(1);
-    expect(runSpy).toHaveBeenCalledWith(2);
-
-    errorSpy.mockRestore();
-    warnSpy.mockRestore();
   });
 });
